backend/requirements.txt
fastapi
uvicorn
motor
pydantic
httpx
python-dotenv
python-multipart
chromadb
langchain
langchain-community
langchain-chroma
langchain-ollama
pymupdf
docx2txt